Renamed everything from pair6 to SkZk
