+++ src/pages/ShoppingListsMain/index.tsx
@@ -12,8 +12,8 @@ const ShoppingListMain = () => {
       <ShoppingListHeader>Our Shopping Lists!</ShoppingListHeader>
       <GroupCardsContainer>
         <ShoppingCard names="name1, name2, name4" onClick={() => history.push(PATHS.EXAMPLE_SHOPPING_PAGE)} />
-        <ShoppingCard names="zhikai, shaokiat" onClick={() => history.push(PATHS.SHOPPING_PAGE)} />
-        <ShoppingCard names="zhikai, shaokiat (Redux)" onClick={() => history.push(PATHS.PAIR6_SHOPPING_PAGE)} />
+        <ShoppingCard names="zhikai, shaokiat" onClick={() => history.push(PATHS.SKZK_SHOPPING_PAGE)} />
+        <ShoppingCard names="zhikai, shaokiat (Redux)" onClick={() => history.push(PATHS.SKZK_SHOPPING_PAGE_REDUX)} />
         <ShoppingCard names="Maxim, Marcus" onClick={() => history.push(PATHS.PAIR_2_SHOPPING_PAGE)} />
         {/* Create group cards here */}
       </GroupCardsContainer>
